Refactored position registry
